move SyncChannel to separate file
